(chore) Restructure globe files
