+++ client/actions/index.js
@@ -1,7 +1,9 @@
 import * as types from '../constants/ActionTypes';
 import helper from '../services/helper';
+import globeHelper from '../services/globe';
 import { push } from 'react-router-redux';
 import { reset } from 'redux-form';
+import planetaryjs from 'planetary.js';
 import axios from 'axios';
 
 // UI Actions
@@ -127,3 +129,66 @@ export const fetchReddit = () => {
   };
 };
 
+// Globe
+
+export const globeInstantiated = () => {
+  return {
+    type: types.GLOBE_INSTANTIATED
+  };
+};
+
+export const instantiateGlobe = () => {
+  const globe = planetaryjs.planet();
+
+  globe.loadPlugin(globeHelper.autorotate(10));
+
+  globe.loadPlugin(planetaryjs.plugins.earth({
+    topojson: { file: 'https://raw.githubusercontent.com/darul75/ng-planetaryjs/master/public/world-110m-withlakes.json' },
+    oceans: { fill: '#000080' },
+    land: { fill: '#339966' },
+    borders: { stroke: '#008000' }
+  }));
+
+  globe.loadPlugin(globeHelper.lakes({
+    fill: '#000080'
+  }));
+
+  globe.loadPlugin(planetaryjs.plugins.pings());
+
+  globe.loadPlugin(planetaryjs.plugins.zoom({
+    scaleExtent: [100, 300]
+  }));
+
+  globe.loadPlugin(planetaryjs.plugins.drag({
+    onDragStart: () => {
+      globe.plugins.autorotate.pause();
+    },
+    onDragEnd: () => {
+      globe.plugins.autorotate.resume();
+    }
+  }));
+
+  globe.projection.scale(175).translate([175, 175]).rotate([0, -10, 0]);
+
+  const colors = ['red', 'yellow', 'white', 'orange', 'green', 'cyan', 'pink'];
+
+  setInterval(() => {
+    const lat = Math.random() * 170 - 85;
+    const lng = Math.random() * 360 - 180;
+    const color = colors[Math.floor(Math.random() * colors.length)];
+    globe.plugins.pings.add(lng, lat, { color, ttl: 2000, angle: Math.random() * 10 });
+  }, 150);
+
+  const canvas = document.getElementById('basicGlobe');
+  canvas.width = 800;
+  canvas.height = 800;
+
+  const context = canvas.getContext('2d');
+  context.scale(2, 2);
+
+  globe.draw(canvas);
+
+  return dispatch => {
+    dispatch(globeInstantiated());
+  };
+};
