+++ client/constants/ActionTypes.js
@@ -3,10 +3,10 @@ export const POPUP_CLOSE = 'POPUP_CLOSE';
 export const POPUP_OPEN = 'POPUP_OPEN';
 
 // Button Actions
-export const REQUEST_POSTS = 'REQUEST_POSTS'
-export const RECEIVE_POSTS = 'RECEIVE_POSTS'
-export const SELECT_SUBREDDIT = 'SELECT_SUBREDDIT'
-export const INVALIDATE_SUBREDDIT = 'INVALIDATE_SUBREDDIT'
+export const REQUEST_POSTS = 'REQUEST_POSTS';
+export const RECEIVE_POSTS = 'RECEIVE_POSTS';
+export const SELECT_SUBREDDIT = 'SELECT_SUBREDDIT';
+export const INVALIDATE_SUBREDDIT = 'INVALIDATE_SUBREDDIT';
 
 export const UPDATE_DATA_STATE = 'UPDATE_DATA_STATE';
 
@@ -17,3 +17,5 @@ export const FLICKR_FAILURE = 'FLICKR_FAILURE';
 export const SHOW_LOCK = 'SHOW_LOCK';
 export const LOCK_SUCCESS = 'LOCK_SUCCESS';
 export const LOCK_ERROR = 'LOCK_ERROR';
+
+export const GLOBE_INSTANTIATED = 'GLOBE_INSTANTIATED';
