+++ client/index.html
@@ -8,12 +8,7 @@
 
   <body>
     <div id="root"></div>
-    <script src="http://planetaryjs.com/css/planetaryjs.css"></script>
-    <script src="http://planetaryjs.com/js/lib/d3.v3.min.js"></script>
-    <script src="http://planetaryjs.com/js/lib/topojson.v1.min.js"></script>
-    <script src="http://planetaryjs.com/js/lib/planetaryjs.min.js"></script>
     <script src="./bundle.js"></script>
-    <script src="http://planetaryjs.com/examples/basic.js"></script>
     <script src="//cdn.auth0.com/js/lock-9.1.min.js"></script>
   </body>
