+++ package.json
@@ -42,7 +42,8 @@
     "redux-form": "^5.2.5",
     "redux-logger": "^2.6.1",
     "redux-promise": "^0.5.3",
-    "redux-thunk": "2.1.0"
+    "redux-thunk": "2.1.0",
+    "topojson": "^1.6.26"
   },
   "devDependencies": {
     "babel": "^6.5.2",
